refactor(ui): Change folder organization for components
