+++ ui/src/forms/AddProtoForm.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import { Field, reduxForm } from 'redux-form';
 import { Row, Col } from 'react-bootstrap';
 
-import { FormInput } from '../components/FormInput/FormInput.js';
-import { TextareaInput } from '../components/FormInput/TextareaInput.js';
-import Button from '../components/CustomButton/CustomButton.js';
+import { FormInput } from '../components/inputs/FormInput.js';
+import { TextareaInput } from '../components/inputs/TextareaInput.js';
+import Button from '../components/themes/CustomButton.js';
 
 let AddProto = props => {
     return (
